requirements.txt
boto3==1.26.65
numpy==1.24.2
pandas==1.5.3
python-dotenv==0.21.1
streamlit==1.17.0
plotly==5.13.0
pytest==7.2.1
